Version1
This update adds the game screen and the main menu. The main menu just has three buttons (only play works right now), and when play is clicked it redirects you to the game_screen. there, there is a random ugly ahh character in the middle with some text above, and two buttons resembling answer choices.